feat: Add transporter and e-invoice fields to sales invoice
- Introduced new columns in `sales_invoice` for transporter details and e-invoice fields.
- Created `e_invoice_responses` table for logging submission attempts and responses.
- Updated backend models and queries to handle new fields.
- Added new API endpoints for fetching transporter branches and submission history.
- Implemented frontend changes to support new fields in invoice forms and display.
- Established data flow for transporter GST and e-invoice submission processes.
- Added tests for new functionality in both backend and frontend.

+++ dbqueries/migrations/add_transporter_einvoice_fields_to_sales_invoice.sql
@@ -4,7 +4,7 @@
 
 -- Add 9 new columns to sales_invoice table
 ALTER TABLE sales_invoice
-ADD COLUMN transporter_branch_id BIGINT NULL AFTER transporter_state_name,
+ADD COLUMN transporter_branch_id INT NULL AFTER transporter_state_name,
 ADD COLUMN transporter_doc_no VARCHAR(255) NULL AFTER transporter_branch_id,
 ADD COLUMN transporter_doc_date DATE NULL AFTER transporter_doc_no,
 ADD COLUMN buyer_order_no VARCHAR(255) NULL AFTER transporter_doc_date,
